feat(dui3): fully reverts helper projects, and cleans up init logic in rhino
