+++ All.sln
@@ -326,15 +326,9 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConnectorRhinoWebUI", "Conn
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "DesktopUI3", "DesktopUI3", "{D29B842B-4AFD-4160-ACFC-01747A9C7BEE}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DUI3WebView2Helper", "DesktopUI3\DUI3WebView2Helper\DUI3WebView2Helper\DUI3WebView2Helper.csproj", "{A5EADC03-1BE3-4B39-9B8B-26CECFB9CB38}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DUI3", "DesktopUI3\DUI3\DUI3.csproj", "{5D437824-4B9D-47DA-903D-70AFC66F0EF9}"
 EndProject
-Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "DUI3CefSharpHelper", "DesktopUI3\DUI3CefSharpHelper\DUI3CefSharpHelper.shproj", "{51A4BD82-8997-4471-8887-A7C82E449652}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DUI3", "DesktopUI3\DUI3\DUI3.csproj", "{633EBB56-B332-45F0-94F1-ECBAABA39C43}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DUI3CefSharp65", "DesktopUI3\DUI3CefSharp65\DUI3CefSharp65.csproj", "{698E458E-88AD-4D20-99E7-14267E058B78}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -1701,22 +1695,6 @@ Global
 		{F0E460C0-ED57-4186-99C4-717194B7ABEE}.Release|Any CPU.Build.0 = Release|Any CPU
 		{F0E460C0-ED57-4186-99C4-717194B7ABEE}.Release|x64.ActiveCfg = Release|Any CPU
 		{F0E460C0-ED57-4186-99C4-717194B7ABEE}.Release|x64.Build.0 = Release|Any CPU
-		{A5EADC03-1BE3-4B39-9B8B-26CECFB9CB38}.Debug Mac|Any CPU.ActiveCfg = Debug|Any CPU
-		{A5EADC03-1BE3-4B39-9B8B-26CECFB9CB38}.Debug Mac|Any CPU.Build.0 = Debug|Any CPU
-		{A5EADC03-1BE3-4B39-9B8B-26CECFB9CB38}.Debug Mac|x64.ActiveCfg = Debug|Any CPU
-		{A5EADC03-1BE3-4B39-9B8B-26CECFB9CB38}.Debug Mac|x64.Build.0 = Debug|Any CPU
-		{A5EADC03-1BE3-4B39-9B8B-26CECFB9CB38}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A5EADC03-1BE3-4B39-9B8B-26CECFB9CB38}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A5EADC03-1BE3-4B39-9B8B-26CECFB9CB38}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{A5EADC03-1BE3-4B39-9B8B-26CECFB9CB38}.Debug|x64.Build.0 = Debug|Any CPU
-		{A5EADC03-1BE3-4B39-9B8B-26CECFB9CB38}.Release Mac|Any CPU.ActiveCfg = Debug|Any CPU
-		{A5EADC03-1BE3-4B39-9B8B-26CECFB9CB38}.Release Mac|Any CPU.Build.0 = Debug|Any CPU
-		{A5EADC03-1BE3-4B39-9B8B-26CECFB9CB38}.Release Mac|x64.ActiveCfg = Debug|Any CPU
-		{A5EADC03-1BE3-4B39-9B8B-26CECFB9CB38}.Release Mac|x64.Build.0 = Debug|Any CPU
-		{A5EADC03-1BE3-4B39-9B8B-26CECFB9CB38}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A5EADC03-1BE3-4B39-9B8B-26CECFB9CB38}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A5EADC03-1BE3-4B39-9B8B-26CECFB9CB38}.Release|x64.ActiveCfg = Release|Any CPU
-		{A5EADC03-1BE3-4B39-9B8B-26CECFB9CB38}.Release|x64.Build.0 = Release|Any CPU
 		{5D437824-4B9D-47DA-903D-70AFC66F0EF9}.Debug Mac|Any CPU.ActiveCfg = Debug|Any CPU
 		{5D437824-4B9D-47DA-903D-70AFC66F0EF9}.Debug Mac|Any CPU.Build.0 = Debug|Any CPU
 		{5D437824-4B9D-47DA-903D-70AFC66F0EF9}.Debug Mac|x64.ActiveCfg = Debug|Any CPU
@@ -1749,22 +1727,6 @@ Global
 		{633EBB56-B332-45F0-94F1-ECBAABA39C43}.Release|Any CPU.Build.0 = Release|Any CPU
 		{633EBB56-B332-45F0-94F1-ECBAABA39C43}.Release|x64.ActiveCfg = Release|Any CPU
 		{633EBB56-B332-45F0-94F1-ECBAABA39C43}.Release|x64.Build.0 = Release|Any CPU
-		{698E458E-88AD-4D20-99E7-14267E058B78}.Debug Mac|Any CPU.ActiveCfg = Debug|Any CPU
-		{698E458E-88AD-4D20-99E7-14267E058B78}.Debug Mac|Any CPU.Build.0 = Debug|Any CPU
-		{698E458E-88AD-4D20-99E7-14267E058B78}.Debug Mac|x64.ActiveCfg = Debug|Any CPU
-		{698E458E-88AD-4D20-99E7-14267E058B78}.Debug Mac|x64.Build.0 = Debug|Any CPU
-		{698E458E-88AD-4D20-99E7-14267E058B78}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{698E458E-88AD-4D20-99E7-14267E058B78}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{698E458E-88AD-4D20-99E7-14267E058B78}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{698E458E-88AD-4D20-99E7-14267E058B78}.Debug|x64.Build.0 = Debug|Any CPU
-		{698E458E-88AD-4D20-99E7-14267E058B78}.Release Mac|Any CPU.ActiveCfg = Debug|Any CPU
-		{698E458E-88AD-4D20-99E7-14267E058B78}.Release Mac|Any CPU.Build.0 = Debug|Any CPU
-		{698E458E-88AD-4D20-99E7-14267E058B78}.Release Mac|x64.ActiveCfg = Debug|Any CPU
-		{698E458E-88AD-4D20-99E7-14267E058B78}.Release Mac|x64.Build.0 = Debug|Any CPU
-		{698E458E-88AD-4D20-99E7-14267E058B78}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{698E458E-88AD-4D20-99E7-14267E058B78}.Release|Any CPU.Build.0 = Release|Any CPU
-		{698E458E-88AD-4D20-99E7-14267E058B78}.Release|x64.ActiveCfg = Release|Any CPU
-		{698E458E-88AD-4D20-99E7-14267E058B78}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -1898,11 +1860,8 @@ Global
 		{B47492D9-2EDA-4016-A930-7FA708C85C3D} = {890F3257-FCC2-4ED8-9180-22B3641B494C}
 		{2D0F9F8A-2E89-4780-978A-CD92D6D7B843} = {BE521908-7944-46F3-98BF-B47D34509934}
 		{F0E460C0-ED57-4186-99C4-717194B7ABEE} = {E94E7327-5A9B-48EE-93CC-E9E9A5B980F1}
-		{A5EADC03-1BE3-4B39-9B8B-26CECFB9CB38} = {D29B842B-4AFD-4160-ACFC-01747A9C7BEE}
 		{5D437824-4B9D-47DA-903D-70AFC66F0EF9} = {D29B842B-4AFD-4160-ACFC-01747A9C7BEE}
-		{51A4BD82-8997-4471-8887-A7C82E449652} = {D29B842B-4AFD-4160-ACFC-01747A9C7BEE}
 		{633EBB56-B332-45F0-94F1-ECBAABA39C43} = {D29B842B-4AFD-4160-ACFC-01747A9C7BEE}
-		{698E458E-88AD-4D20-99E7-14267E058B78} = {D29B842B-4AFD-4160-ACFC-01747A9C7BEE}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {1D43D91B-4F01-4A78-8250-CC6F9BD93A14}
