+++ ConnectorRhino/ConnectorRhino.slnf
@@ -17,9 +17,6 @@
       "DesktopUI2\\AvaloniaHwndHost\\AvaloniaHwndHost.csproj",
       "DesktopUI2\\DesktopUI2\\DesktopUI2.csproj",
       "DesktopUI3\\DUI3\\DUI3.csproj",
-      "DesktopUI3\\DUI3CefSharp65\\DUI3CefSharp65.csproj",
-      "DesktopUI3\\DUI3CefSharpHelper\\DUI3CefSharpHelper.shproj",
-      "DesktopUI3\\DUI3WebView2Helper\\DUI3WebView2Helper\\DUI3WebView2Helper.csproj",
       "Objects\\Converters\\ConverterRhinoGh\\ConverterGrasshopper6\\ConverterGrasshopper6.csproj",
       "Objects\\Converters\\ConverterRhinoGh\\ConverterGrasshopper7\\ConverterGrasshopper7.csproj",
       "Objects\\Converters\\ConverterRhinoGh\\ConverterRhino6\\ConverterRhino6.csproj",
