--- ConnectorRhino/ConnectorRhinoWebUI/SpeckleWebUIPanelCef.xaml
@@ -1,16 +1,0 @@
-﻿<UserControl x:Class="ConnectorRhinoWebUI.SpeckleWebUIPanelCef"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:ConnectorRhinoWebUI" 
-             xmlns:cefSharp="clr-namespace:CefSharp.Wpf;assembly=CefSharp.Wpf"
-      mc:Ignorable="d" 
-             d:DesignHeight="450" d:DesignWidth="800">
-  <Grid>
-    <cefSharp:ChromiumWebBrowser
-      Name="Browser"
-      Grid.Row="0"
-      Address="http://localhost:8082" />
-  </Grid>
-</UserControl>
