+++ ConnectorRhino/ConnectorRhinoWebUI/WebUIPanelWebView2.xaml
@@ -1,8 +1,9 @@
-﻿<UserControl x:Class="DUI3WebView2Helper.WebView2UserControl"
+﻿<UserControl x:Class="ConnectorRhinoWebUI.WebUiPanelWebView2"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:ConnectorRhinoWebUI" 
              xmlns:wv2="clr-namespace:Microsoft.Web.WebView2.Wpf;assembly=Microsoft.Web.WebView2.Wpf"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
