--- DesktopUI3/DUI3CefSharp65/packages.config
@@ -1,7 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="CefSharp.Common" version="65.0.1" targetFramework="net471" />
-  <package id="CefSharp.Wpf" version="65.0.1" targetFramework="net471" />
-  <package id="cef.redist.x64" version="3.3325.1758" targetFramework="net471" />
-  <package id="cef.redist.x86" version="3.3325.1758" targetFramework="net471" />
-</packages>
