--- DesktopUI3/DUI3CefSharpHelper/CefSharpUserControl.xaml
@@ -1,12 +1,0 @@
-﻿<UserControl x:Class="DUI3CefSharpHelper.CefSharpUserControl"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             mc:Ignorable="d"
-             d:DesignHeight="300" d:DesignWidth="300">
-  <Grid>
-    
-  </Grid>
-</UserControl>
-
