--- DesktopUI3/DUI3CefSharpHelper/CefSharpUserControl.xaml.cs
@@ -1,12 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace DUI3CefSharpHelper;
-
-public partial class CefSharpUserControl : UserControl
-{
-  public CefSharpUserControl()
-  {
-    InitializeComponent();
-  }
-}
-
