--- DesktopUI3/DUI3CefSharpHelper/DUI3CefSharpHelper.projitems
@@ -1,20 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
-    <HasSharedItems>true</HasSharedItems>
-    <SharedGUID>E688D82F-37EB-4E83-8891-D2307FB4AC1F</SharedGUID>
-  </PropertyGroup>
-  <PropertyGroup Label="Configuration">
-    <Import_RootNamespace>DUI3CefSharpHelper</Import_RootNamespace>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)CefSharpHelper.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)CefSharpUserControl.xaml.cs">
-      <DependentUpon>CefSharpUserControl.xaml</DependentUpon>
-    </Compile>
-  </ItemGroup>
-  <ItemGroup>
-    <Page Include="$(MSBuildThisFileDirectory)CefSharpUserControl.xaml" />
-  </ItemGroup>
-</Project>
