--- DesktopUI3/DUI3WebView2Helper/DUI3WebView2Helper/WebView2UserControl.xaml.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Windows.Controls;
-
-namespace DUI3WebView2Helper;
-
-public partial class WebView2UserControl : UserControl
-{
-  public WebView2UserControl()
-  {
-    InitializeComponent();
-  }
-}
-
