--- DesktopUI3/DUI3WebView2Helper/DUI3WebView2Helper/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.Web.WebView2" version="1.0.1901.177" targetFramework="net472" />
-</packages>
